integrated astro and tailwind
we want to use astro because astro is a great framework for static sites
like in our use case (hackathon site). its mpa and we aren't sharing any
data between different pages so we are chilling. the parts that are
static, which is a lot, this is great because we aren't shipping js to
the browser for those parts which makes the site much faster. the parts
that are more interactive, like some dashboard components for hackers,
can use server islands. oh and astro has great seo components that we
can use

https://github.com/jonasmerlin/astro-seo