implemented react query | showcase the jwt token in dashboard if logged in
as title. we are using queryClient and passing it in to every react
query function because astro is a mpa and in order to implement the
query provider, we would need to do some hacks and convert astro into a
spa which might reduce the performantness of astro itself.

removed query wrapper